+++ assets/js/app.js
@@ -1,8 +1,10 @@
 var submit = document.getElementById('transform_qr');
 var url = document.getElementById('url');
+var qr = new QRCode(
+  document.getElementById("qrcode")
+);
+
 submit.addEventListener('click', event => {
-  new QRCode(
-    document.getElementById("qrcode"),
-    url.value
-  );
+  qr.clear();
+  qr.makeCode(url.value);
 });