+++ CHANGELOG.md
@@ -3,9 +3,10 @@
 
 ### Bug Fixes
 
-* add jquery and qrcode ([6bc193b](https://github.com/citosid/qr/commit/6bc193be8055814ae599864a26a27e48a6bfe20e))
+* add jquery and qrcode ([c94a47f](https://github.com/citosid/qr/commit/c94a47fac48a7e4c76ab614c69804727b8fc8b15))
 * add script at the end so it has the html loaded ([df8ac34](https://github.com/citosid/qr/commit/df8ac343ec1ff19bb311f6db5b03722c5fd2c2da))
 * fix path for js assets ([1c188d4](https://github.com/citosid/qr/commit/1c188d491d17b3e77ec7073ecc0122270f2e0c17))
+* the code should be created only once ([86f6e0b](https://github.com/citosid/qr/commit/86f6e0bc0cfb2e4455b4f5d8468f05d4e6e0fa31))
 
 
 ### Features